Support exec and add basic logic implementations
This adds `exec(Vec<AstNode>)` to create and execute a program. It also
adds some basic standard library functions - namely `if`, `eq?` and
`neq?`. Likely `neq?` could be implemented in pure risk if `not` is
implemented. (`(define (neq x y) (not (eq? x y)))`, however, definition
is not supported yet so a todo has been added.