Support definition, fix calling AstNode::Leaf bug (#6)
This adds support for definition of variabled and functions via the
form `(define var expr)` and `(define (func ...params) body)`

It also fixes a bug where Primitive::S values would call themselves even
if not in a function execution context (`if` vs `(if)`). Now `ref`
returns the reference it is and execution is gated by parens as it
should be.